chore: Add `timeoutedPromise` to all tests in Example
